+++ src/libponyc/codegen/genjit.cc
@@ -2,20 +2,159 @@
 #include "genexe.h"
 #include "genopt.h"
 
+#if PONY_LLVM >= 700
+
 #include "llvm_config_begin.h"
-
-#include <llvm/IR/Mangler.h>
-#include <llvm/IR/Module.h>
-#include <llvm/ExecutionEngine/Orc/CompileUtils.h>
-#include <llvm/ExecutionEngine/Orc/IRCompileLayer.h>
-#include <llvm/ExecutionEngine/Orc/LambdaResolver.h>
-#if PONY_LLVM >= 500
+#  include <llvm/ExecutionEngine/ExecutionEngine.h>
+#  include <llvm/ExecutionEngine/JITSymbol.h>
+#  include <llvm/ExecutionEngine/SectionMemoryManager.h>
+#  include <llvm/ExecutionEngine/Orc/CompileUtils.h>
+#  include <llvm/ExecutionEngine/Orc/Core.h>
+#  include <llvm/ExecutionEngine/Orc/Legacy.h>
+#  include <llvm/ExecutionEngine/Orc/IRCompileLayer.h>
 #  include <llvm/ExecutionEngine/Orc/RTDyldObjectLinkingLayer.h>
-#else
-#  include <llvm/ExecutionEngine/Orc/ObjectLinkingLayer.h>
-#endif
-#include <llvm/ExecutionEngine/SectionMemoryManager.h>
+#  include <llvm/IR/DataLayout.h>
+#  include <llvm/IR/Mangler.h>
+#  include <llvm/Support/Error.h>
+#  include <llvm/Target/TargetMachine.h>
+#include "llvm_config_end.h"
 
+using namespace llvm;
+using namespace llvm::orc;
+
+class PonyJIT
+{
+  ExecutionSession _es;
+  std::shared_ptr<SymbolResolver> _sr;
+  std::unique_ptr<TargetMachine> _tm;
+  const DataLayout _dl;
+  RTDyldObjectLinkingLayer _obj_layer;
+  IRCompileLayer<decltype(_obj_layer), SimpleCompiler> _compile_layer;
+
+public:
+  Error error;
+
+  PonyJIT() :
+    _es(),
+    _sr(createLegacyLookupResolver(_es,
+      [this](const std::string& name) -> JITSymbol
+      {
+        auto symbol = _compile_layer.findSymbol(name, false);
+        if (symbol) return symbol;
+
+        auto err = symbol.takeError();
+        if (err) return std::move(err);
+
+        auto symaddr = RTDyldMemoryManager::getSymbolAddressInProcess(name);
+        if (symaddr)
+          return JITSymbol(symaddr, JITSymbolFlags::Exported);
+
+        return nullptr;
+      },
+      [this](Error err)
+      {
+        error = std::move(err);
+      })),
+    _tm(EngineBuilder().selectTarget()),
+    _dl(_tm->createDataLayout()),
+    _obj_layer(_es,
+      [this](VModuleKey)
+      {
+        return RTDyldObjectLinkingLayer::Resources
+        {
+          std::make_shared<SectionMemoryManager>(), _sr
+        };
+      }),
+    _compile_layer(_obj_layer, SimpleCompiler(*_tm)),
+    error(Error::success())
+  {
+    llvm::sys::DynamicLibrary::LoadLibraryPermanently(nullptr);
+  }
+
+  VModuleKey addModule(std::unique_ptr<Module> module)
+  {
+    VModuleKey key = _es.allocateVModule();
+    cantFail(_compile_layer.addModule(key, std::move(module)));
+    return key;
+  }
+
+  JITTargetAddress getSymbolAddress(const std::string& name)
+  {
+    std::string mangled;
+    raw_string_ostream mangled_stream(mangled);
+    Mangler::getNameWithPrefix(mangled_stream, name, _dl);
+    JITSymbol symbol = _compile_layer.findSymbol(mangled_stream.str(), false);
+    return cantFail(symbol.getAddress());
+  }
+};
+
+bool gen_jit_and_run(compile_t* c, int* exit_code, jit_symbol_t* symbols,
+  size_t symbol_count)
+{
+  reach_type_t* t_main = reach_type_name(c->reach, "Main");
+  reach_type_t* t_env = reach_type_name(c->reach, "Env");
+
+  if ((t_main == NULL) || (t_env == NULL))
+    return false;
+
+  gen_main(c, t_main, t_env);
+
+  if (!genopt(c, true))
+    return false;
+
+  if (LLVMVerifyModule(c->module, LLVMReturnStatusAction, NULL) != 0)
+    return false;
+
+  PonyJIT jit;
+  jit.addModule(std::unique_ptr<Module>(unwrap(c->module)));
+  c->module = nullptr;
+
+  if (jit.error)
+  {
+    errorf(c->opt->check.errors, nullptr, "LLVM ORC JIT Error");
+    return false;
+  }
+
+  for (size_t i = 0; i < symbol_count; i++)
+  {
+    void* address = (void*)jit.getSymbolAddress(symbols[i].name);
+    if (address == nullptr)
+      return false;
+
+    memcpy(symbols[i].address, address, symbols[i].size);
+  }
+
+  auto main = reinterpret_cast<int(*)(int, const char**, const char**)>(
+    jit.getSymbolAddress("main")
+  );
+
+  if (main == nullptr)
+    return false;
+
+  const char* argv[] = { "ponyjit", nullptr };
+  const char* envp = nullptr;
+
+  int ec = main(1, argv, &envp);
+
+  if (exit_code != nullptr)
+    *exit_code = ec;
+  return true;
+}
+
+#else
+
+#include "llvm_config_begin.h"
+#  include <llvm/IR/Mangler.h>
+#  include <llvm/IR/Module.h>
+#  include <llvm/ExecutionEngine/Orc/CompileUtils.h>
+#  include <llvm/ExecutionEngine/Orc/IRCompileLayer.h>
+#  include <llvm/ExecutionEngine/Orc/LambdaResolver.h>
+#  if PONY_LLVM >= 500
+#    include <llvm/ExecutionEngine/Orc/RTDyldObjectLinkingLayer.h>
+#  else
+#    include <llvm/ExecutionEngine/Orc/ObjectLinkingLayer.h>
+#  endif
+#  include <llvm/ExecutionEngine/SectionMemoryManager.h>
 #include "llvm_config_end.h"
 
 namespace orc = llvm::orc;
@@ -76,15 +215,15 @@ bool gen_jit_and_run(compile_t* c, int* exit_code, jit_symbol_t* symbols,
   reach_type_t* t_main = reach_type_name(c->reach, "Main");
   reach_type_t* t_env = reach_type_name(c->reach, "Env");
 
-  if((t_main == NULL) || (t_env == NULL))
+  if ((t_main == NULL) || (t_env == NULL))
     return false;
 
   gen_main(c, t_main, t_env);
 
-  if(!genopt(c, true))
+  if (!genopt(c, true))
     return false;
 
-  if(LLVMVerifyModule(c->module, LLVMReturnStatusAction, NULL) != 0)
+  if (LLVMVerifyModule(c->module, LLVMReturnStatusAction, NULL) != 0)
     return false;
 
   // The Orc JIT wants the module in a shared_ptr, but we don't want to transfer
@@ -102,6 +241,7 @@ bool gen_jit_and_run(compile_t* c, int* exit_code, jit_symbol_t* symbols,
     [&mem_mgr]{ return mem_mgr; }
 #endif
   };
+
   CompileLayerType compile_layer{linking_layer, orc::SimpleCompiler{*machine}};
 
   auto local_lookup = [&compile_layer](llvm::StringRef name)
@@ -181,3 +321,5 @@ bool gen_jit_and_run(compile_t* c, int* exit_code, jit_symbol_t* symbols,
 
   return true;
 }
+
+#endif