+++ src/libponyc/codegen/genreference.c
@@ -112,17 +112,6 @@ LLVMValueRef gen_fieldload(compile_t* c, ast_t* ast)
 
   field = LLVMBuildLoad(c->builder, field, "");
 
-  type = deferred_reify(reify, ast_type(left), c->opt);
-  LLVMValueRef metadata = tbaa_metadata_for_type(c, type);
-  ast_free_unattached(type);
-
-#if PONY_LLVM >= 400
-  tbaa_tag(c, metadata, field);
-#else
-  const char id[] = "tbaa";
-  LLVMSetMetadata(field, LLVMGetMDKindID(id, sizeof(id) - 1), metadata);
-#endif
-
   return gen_assign_cast(c, c_t->use_type, field, t->ast_cap);
 }
 
@@ -263,8 +252,16 @@ LLVMValueRef gen_localdecl(compile_t* c, ast_t* ast)
   LLVMMetadataRef file = codegen_difile(c);
   LLVMMetadataRef scope = codegen_discope(c);
 
+#if PONY_LLVM >= 700
+  uint32_t align_bytes = LLVMABIAlignmentOfType(c->target_data, c_t->mem_type);
+
+  LLVMMetadataRef info = LLVMDIBuilderCreateAutoVariable(c->di, scope,
+    name, strlen(name), file, (unsigned)ast_line(ast), c_t->di_type,
+    true, LLVMDIFlagZero, align_bytes * 8);
+#else
   LLVMMetadataRef info = LLVMDIBuilderCreateAutoVariable(c->di, scope, name,
     file, (unsigned)ast_line(ast), c_t->di_type);
+#endif
 
   LLVMMetadataRef expr = LLVMDIBuilderCreateExpression(c->di, NULL, 0);